Ebrahim/src/test/java/com/aharpour/adventofcode/Day7Question1Test.java
package com.aharpour.adventofcode;

import org.junit.Test;


import static org.junit.Assert.assertEquals;

public class Day7Question1Test {

    private Day7Question1 question = new Day7Question1();

    @Test
    public void givenCases() {
        assertEquals("CABDFE", question.calculate("Step C must be finished before step A can begin.\n" +
                "Step C must be finished before step F can begin.\n" +
                "Step A must be finished before step B can begin.\n" +
                "Step A must be finished before step D can begin.\n" +
                "Step B must be finished before step E can begin.\n" +
                "Step D must be finished before step E can begin.\n" +
                "Step F must be finished before step E can begin."));
    }
}